draw_sprite

+++ sprite.h
@@ -1,2 +1,2 @@
-void draw_sprite(int left, int top, int width, int height, int centre_x, int centre_y, float scale,
+void draw_sprite(int left, int top, int width, int height, float centre_x, float centre_y, float scale,
 	    struct texture *tex);
